requirements.txt
einops>=0.8.0
fast_pytorch_kmeans>=0.2.0.1
icecream>=2.1.3
ipython>=8.10.0
lpips>=0.1.4
matplotlib>=3.7.0
numpy>=1.23.5
opencv_python>=4.9.0.80
opencv_python_headless>=4.9.0.80
Pillow>=10.3.0
pytorch_msssim>=0.2.1
PyYAML>=6.0.1
scikit_image>=0.19.3
scikit_learn>=1.2.1
scipy>=1.10.0
timm>=0.9.2
TorchSnooper>=0.2
torchsummary>=1.5.1
tqdm>=4.64.1
Wand>=0.6.11
wandb>=0.15.3
